requirements.txt
absl-py==0.11.0
astor==0.8.1
autocorrect==2.3.0
bleach==1.5.0
click==7.1.2
editdistance==0.5.3
Flask==1.1.2
gast==0.4.0
grpcio==1.35.0
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
protobuf==3.14.0
pytesseract==0.3.7
six==1.15.0
tensorboard==1.8.0
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.8.0-py3-none-any.whl
termcolor==1.1.0
tesseract==0.1.3
Werkzeug==1.0.1
